DP

Description

This module defines a class of non-deterministic dynamic programming problems and the means to solve/optimize them. The framework targets a slightly different set of problems, but some parts take after: https://hackage.haskell.org/package/HyloDP-1.0.0/docs/HyloDP-Base.html, but targets

This module does: - provide modeling for non-deterministic dynamic programming - this allows the "reward" of an action to depend on multiple successor states - abstract over the search procedure used to determine the optimal action, allowing for: - exhaustive searches over finite domains - a golden section search for unimodal objectives over "infinite" domains - something smarter if you want? - solve the problem while returning the entire table of subsolutions - (the solver needs the entire table anyway).

It does not: - use algebraic infrastructure to explain dynamic programming is actually described by a hylomorphism (not sure if this one actually still is) - use MemoTrie (it explodes for some reason, but generally speaking the hashmap isn't the bottleneck anyway) - use Semirings to describe formulate optimality and the reported statistics at the type level

Documentation

data DPProblem s v a m

The type of dynamic programming problems valued in v, with states s and actions a.

The (usually) Monad m can be used to compute something if you need, usually it's one of: - Identity - IO/ST - a transformer stack including memoization

(The decomposition into subproblems is encoded in actValue).

Constructors

DPProblem

Fields

search :: Search s v a m

The search procedure over a.

actValue :: s -> (s -> m v) -> a -> m v

Returns the objective value of performing an action a in a state s.

enumSearch :: (Ord v, Monad m) => (s -> Either v [a]) -> Search s v a m

Given a function `s -> Either v [a]` indicating whether a state is either: - terminal with value v - allows for actions `[a]` to be taken this procedure exhaustively searches the list for the action with the optimal objective.

uniSearch :: (Ord v, Monad m, Floating x) => Int -> (x -> a) -> x -> x -> Search s v a m

Given an injection mk of values x into actions a, this procedure performs a golden section search over x for the objective of `mk x`.

execDP :: Hashable s => DPProblem s v a IO -> s -> IO (HashMap s (v, Maybe a))

execDP solves a DPProblem in IO from a given initial state, returning the entire table mapping states to their optimal actions and objective values. Uses an IORef to a HashMap for memoization.

reducedP :: (Monad m, Ord v, Hashable s) => HashMap s (v, Maybe a) -> DPProblem s v a m -> DPProblem s v a m

A "trick" to reduce the output of execDP to only the relevant parts. Instantiate this with the output of an execDP call and rerun execDP with this problem instead.

reducedDP :: (Hashable s, Ord v) => DPProblem s v a IO -> s -> IO (HashMap s (v, Maybe a))

reducedDP solves a DPProblem in IO from a given initial state, returning the table mapping reachable states to their optimal actions and objective values.

A state s is reachable if it's either initial, or the objective value of a reachable state s' after taking the optimal action refers to s.