Probability

Description

This module provides some wrappers for computing some statistics for your DPProblem.

Make sure your inputs are independent if you like having correct outputs.

Documentation

avg :: Mixture a => [a] -> a

Compute the value of the statistic for a uniform mixture of distributions.

class HasMean a r | a -> r where

The class of statistics containing enough informations to calculate an expected value.

Remark, if `HasMean a r`, `Mixture a`, and `Mixture r`, it does not necessarily hold that `mean . weightedAvg == weightedAvg . fmap (second mean)`. For example, see Log1P.

Methods

mean :: a -> r

Extract the expected value. Generally this should satisfy `mean . fromRational = id`.

class HasVariance a r | a -> r where

Methods

variance :: a -> r

class Mixture a where

A statistic `V :: X -> v` supports mixtures if: - there is a function $m$ - $f_X(x) = pf_Y(x) + (1-p)f_Z(x)$ implies $V[X] = m(p, V[Y], V[Z])

The method weightedAvg generalizes this to the $n$-ary case.

Methods

weightedAvg :: [(Double, a)] -> a

Compute the value of the statistic for a mixture of distributions.

newtype Log1P a

A wrapper to compute the expected value of the logarithm of X plus one, E[log(1 + X)]. This is equivalent to the logarithm of the expected geometric growth rate.

Note that all this does is map an incoming value x to `log (1 + x)` and outgoing values x to `exp(x) - 1`.

Constructors

Log1P

Fields

getLog1P :: a

newtype Mean a

A wrapper to compute the expected value E[X], also known as the Identity.

Constructors

Mean

Fields

getMean :: a

newtype MaxMean a

A wrapper to search for the objective value with the maximal expected value.

This ensures that if `mean x < mean y` then `MaxMean x < MaxMean y`.

Constructors

MaxMean

Fields

getMaxMean :: a

newtype MinMean a

A wrapper to search for the objective value with the minimal expected value.

This ensures that if `mean x >= mean y` then `MinMean x < MinMean y`.

Constructors

MinMean

Fields

getMinMean :: a

data Variance a b

A wrapper to compute the variance and expected value of X. Requires its first argument to have expected values to work.

Constructors

Variance

Fields

unVariance :: !a

getVariance :: !b